E preparation system Real Time Quizz
In this application user able  to attend Quizz test real time test. that shows the real time score of an users. they will provide mcq types questions to user  for there particular subjects. there is also role in admin which have the liberty to add quizz and subjects that he or she wanted.In this application is specially created for those students who are preparating for exams.